Added tests to opencv_io
