+++ tests/test_camera/testscamera.py
@@ -1,0 +1,1 @@
+from labvision import camera
